cleanup tolerance cli test
