+++ tests/cli/test_stats_cli.py
@@ -10,6 +10,7 @@ from tests.helpers.fixtures import (  # noqa: F401
     tmp_dir,
 )
 from tests.helpers.helpers import (
+    assert_empty_df,
     load_pandas,
     pandas_error,
     run_perturb_cli,
@@ -28,7 +29,7 @@ def test_stats_cli_no_ensemble(nc_with_T_U_V, df_ref_stats):
     df_test = load_pandas(stats_file, index_col=[0, 1, 2], header=[0, 1])
     err = pandas_error(df_ref_stats, df_test)
 
-    assert len(err.values) == 0, "Stats datasets are not equal!\n{}".format(err)
+    assert_empty_df(err, "Stats datasets are not equal!")
 
 
 @pytest.mark.xfail(
@@ -52,7 +53,7 @@ def test_stats_cli_ensemble_with_too_small_perturb_amplitude_for_member(
     )
     err = pandas_error(df_ref_ensemble_stats[member], df_test)
 
-    assert len(err.values) == 0, "Stats datasets are not equal!\n{}".format(err)
+    assert_empty_df(err, "Stats datasets are not equal!")
 
 
 @pytest.mark.parametrize("member", range(1, 11))
@@ -71,4 +72,4 @@ def test_stats_cli_ensemble_for_member(nc_with_T_U_V, df_ref_ensemble_stats, mem
     )
     err = pandas_error(df_ref_ensemble_stats[member], df_test)
 
-    assert len(err.values) == 0, "Stats datasets are not equal!\n{}".format(err)
+    assert_empty_df(err, "Stats datasets are not equal!")
